Changed view and image clear cache exclude index.html

+++ shift/system/library/image.php
@@ -128,6 +128,10 @@ class Image extends SimpleImage
             $nodes = new \RecursiveIteratorIterator($dirIterator, \RecursiveIteratorIterator::CHILD_FIRST);
 
             foreach ($nodes as $node) {
+                if ($node->getBasename() === 'index.html') {
+                    continue;
+                }
+
                 $node->isDir() ? rmdir($node->getRealPath()) : unlink($node->getRealPath());
             }
         }
